LegalEase_Aayush_FullStack/backend/requirements.txt
# LegalEase AI – Python Backend Dependencies

# Web Framework
fastapi==0.115.0
uvicorn[standard]==0.30.6

# HTTP Client (for Gemini API calls)
httpx==0.27.2

# PDF Generation
reportlab==4.2.2

# QR Code
qrcode[pil]==7.4.2
Pillow==10.4.0

# Data Validation
pydantic==2.9.2
pydantic-settings==2.5.2

# Environment Variables
python-dotenv==1.0.1

# CORS (included in fastapi)
# SQLite (included in Python stdlib)
